WIP: Add cydia installation, ssh and launch daemons. **THIS COMMIT IS NOT IN A WORKING STATE** The AMFI bypass is very hacky, and is not scalable for our needs. This code will work when the amfi bypass is updated to be better, as stated in the README of upstream.
The code has been tested on my device, and it extracts the files
correctly, but cydia doesn't show on the homescreen. The cache changed
in iOS 11, and this will need more looking into.

The code has also been split up into neater groups, and some parts of
the code refactored, and some new blocks created.